Added the sentry logger for gogstash function within same function
